Allow defining complete scoped_name with modules
