+++ dds_gen/src/generator/rust.rs
@@ -865,12 +865,7 @@ impl<'a> RustGenerator<'a> {
     }
 
     fn scoped_name(&mut self, pair: IdlPair) {
-        let identifier = pair
-            .into_inner()
-            .next()
-            .expect("Must have an identifier according to the grammar");
-
-        self.writer.push_str(identifier.as_str())
+        self.writer.push_str(pair.as_str())
     }
 
     fn const_dcl(&mut self, pair: IdlPair) {
